refactor to seperate component files
